add ProductService , ShareholderBrandService

+++ src/main/java/service/BrandService.java
@@ -2,8 +2,8 @@ package service;
 
 import model.Brand;
 import repository.BrandRepository;
+import utility.ApplicationContex;
 import utility.Validation;
-
 import java.sql.SQLException;
 import java.util.Scanner;
 
@@ -11,6 +11,8 @@ public class BrandService {
 
     private final Scanner scanner = new Scanner(System.in);
     private final BrandRepository brandRepository;
+    private final ProductService productService = ApplicationContex.getProductService();
+    private final ShareholderBrandService shareholderBrandService = ApplicationContex.getShareholderBrandService();
 
     public BrandService(BrandRepository brandRepository) {
         this.brandRepository = brandRepository;
@@ -33,68 +35,77 @@ public class BrandService {
                 System.out.println("please enter correct website");
         }
 
-            System.out.println("Please enter a description of your brand: ");
-            String description = scanner.next();
+        System.out.println("Please enter a description of your brand: ");
+        String description = scanner.next();
 
-            Brand brand = new Brand(name, website, description);
+        Brand brand = new Brand(name, website, description);
 
-            int result = brandRepository.CreateBrand(brand);
+        int result = brandRepository.CreateBrand(brand);
 
-            if (result == 1)
-                System.out.println(name + "you successfully added :)");
-            else
-                System.out.println("something is wrong :/");
+        if (result == 1)
+            System.out.println(name + "you successfully added :)");
+        else
+            System.out.println("something is wrong :/");
 
-        }
+    }
 
-        public void editBrand () throws SQLException {
+    public void editBrand() throws SQLException {
 
-            System.out.println("please enter your Brand Id: ");
-            int id = scanner.nextInt();
+        System.out.println("please enter your Brand Id: ");
+        int id = scanner.nextInt();
 
-            Brand brand = load(id);
+        Brand brand = load(id);
 
-            System.out.println("please enter your Brand name: ");
-            String BrandName = scanner.next();
+        System.out.println("please enter your Brand name: ");
+        String BrandName = scanner.next();
 
-            brand.setBrandName(BrandName);
+        brand.setBrandName(BrandName);
 
-            System.out.println("please enter your Brand website: ");
-            String WebsiteBrand = scanner.next();
+        System.out.println("please enter your Brand website: ");
+        String WebsiteBrand = scanner.next();
 
-            brand.setWebsite(WebsiteBrand);
+        brand.setWebsite(WebsiteBrand);
 
-            System.out.println("please enter your Brand description: ");
-            String BrandDis = scanner.next();
+        System.out.println("please enter your Brand description: ");
+        String BrandDis = scanner.next();
 
-            brand.setDescription(BrandDis);
+        brand.setDescription(BrandDis);
 
-            brandRepository.editBrand(brand);
-            System.out.println(brand);
+        brandRepository.editBrand(brand);
+        System.out.println(brand);
 
 //        if (result == 1)
 //            System.out.println( BrandName + "you successfully edited :)");
 //        else
 //            System.out.println("something is wrong :/");
 
-        }
-
-        private Brand load ( int id) throws SQLException {
-
-            Brand brand = brandRepository.load(id);
-            return brand;
-        }
-
-        public void deleteBrand () throws SQLException {
-
-            System.out.println("enter brandId that you want delete :");
-            int id = scanner.nextInt();
-
-            Brand brand = load(id);
-            brandRepository.deleteBrand(brand);
-
-
-        }
     }
+
+    private Brand load(int id) throws SQLException {
+
+        Brand brand = brandRepository.load(id);
+        return brand;
+    }
+
+    public void deleteBrand() throws SQLException {
+
+        System.out.println("enter brandId that you want delete :");
+        int id = scanner.nextInt();
+
+        Brand brand = load(id);
+        if (brand != null) {
+
+
+            productService.deleteBrandId(id);
+            shareholderBrandService.deleteBrandId(id);
+            brandRepository.deleteBrand(brand);
+            System.out.println("Brand deleted successfully");
+        } else {
+            System.out.println("Brand not found");
+        }
+
+
+    }
+}
